+++ lib/claude_on_rails/railtie.rb
@@ -3,10 +3,15 @@
 require 'rails/railtie'
 
 module ClaudeOnRails
-  # Railtie to load rake tasks in Rails applications
+  # Railtie to load rake tasks and generators in Rails applications
   class Railtie < Rails::Railtie
     rake_tasks do
       load File.expand_path('../tasks/claude_on_rails.rake', __dir__)
+    end
+
+    generators do
+      require 'generators/claude_on_rails/swarm/swarm_generator'
+      require 'generators/claude_on_rails/subagents/subagents_generator'
     end
   end
 end